Split codesize tests into separate test file. NFC
This makes it easier to run all the code size tests.

test_other.py is currently 16,000 lines so splitting this out is good
to that problem too.
