+++ test/runner.py
@@ -398,7 +398,7 @@ def flattened_tests(loaded_tests):
 
 
 def suite_for_module(module, tests, options):
-  suite_supported = module.__name__ in ('test_core', 'test_other', 'test_posixtest', 'test_browser')
+  suite_supported = module.__name__ in ('test_core', 'test_other', 'test_posixtest', 'test_browser', 'test_codesize')
   if not common.EMTEST_SAVE_DIR and not shared.DEBUG:
     has_multiple_tests = len(tests) > 1
     has_multiple_cores = parallel_testsuite.num_cores() > 1
